+++ main_hm/sdl2.c
@@ -101,7 +101,7 @@ void initFramerate_SDL() {
     SDL_initFramerate(&fpsm);
 }
 
-void setFramerate_SDL(Uint32 rate) {
+void setFramerate_SDL(float rate) {
     if (SDL_setFramerate(&fpsm,rate) < 0) {
         printf("SDL_glx: Couldn't set frame rate\n");
         SDL_Quit();
